minimap in map resizing
